Project restructure.
- Moved pages folder into src folder.
- Moved utils folder into src folder.
