+++ src/pages/trades/TradesAnalysisPage.tsx
